shiny/requirements.txt
shiny
pandas
plotly
shinywidgets
snowflake-connector-python
posit-sdk
